backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.128.0
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
joblib==1.5.3
numpy==2.4.1
openpyxl==3.1.5
pandas==2.3.3
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
